add enum next/previous
